Extract Endpath into the extended library
