+++ libs/extended/extended.cabal
@@ -23,6 +23,7 @@ library
     Network.RabbitMqAdmin
     Options.Applicative.Extended
     Servant.API.Extended
+    Servant.API.Extended.Endpath
     Servant.API.Extended.RawM
     System.Logger.Extended
